added testnet

+++ packages/controllers/install.go
@@ -27,6 +27,7 @@ import (
 	"github.com/EGaaS/go-egaas-mvp/packages/static"
 	"github.com/EGaaS/go-egaas-mvp/packages/utils"
 	"github.com/astaxie/beego/config"
+	"strings"
 )
 
 func (c *Controller) Install() (string, error) {
@@ -40,11 +41,7 @@ func (c *Controller) Install() (string, error) {
 	if generateFirstBlock != "" {
 		*utils.GenerateFirstBlock = utils.StrToInt64(generateFirstBlock)
 	}
-	firstBlockDir := c.r.FormValue("first_block_dir")
-	*utils.FirstBlockDir = *utils.Dir
-	if firstBlockDir != "" {
-		*utils.FirstBlockDir = firstBlockDir
-	}
+
 	installType := c.r.FormValue("type")
 	tcpHost := c.r.FormValue("tcp_host")
 	if tcpHost != "" {
@@ -58,7 +55,7 @@ func (c *Controller) Install() (string, error) {
 	if logLevel != "DEBUG" {
 		logLevel = "ERROR"
 	}
-	url := c.r.FormValue("url")
+	first_load_blockchain_url := c.r.FormValue("first_load_blockchain_url")
 	firstLoad := c.r.FormValue("first_load")
 	dbType := c.r.FormValue("db_type")
 	dbHost := c.r.FormValue("host")
@@ -66,9 +63,17 @@ func (c *Controller) Install() (string, error) {
 	dbName := c.r.FormValue("db_name")
 	dbUsername := c.r.FormValue("username")
 	dbPassword := c.r.FormValue("password")
+	firstBlockDir := c.r.FormValue("first_block_dir")
 
-	if len(url) == 0 {
-		url = consts.BLOCKCHAIN_URL
+	if firstLoad == `Private-net` {
+		*utils.FirstBlockDir = *utils.Dir
+		if firstBlockDir != "" {
+			*utils.FirstBlockDir = firstBlockDir
+		}
+	}
+
+	if len(first_load_blockchain_url) == 0 {
+		first_load_blockchain_url = consts.BLOCKCHAIN_URL
 	}
 
 	if _, err := os.Stat(*utils.Dir + "/config.ini"); os.IsNotExist(err) {
@@ -150,7 +155,7 @@ func (c *Controller) Install() (string, error) {
 		return "", utils.ErrInfo(err)
 	}
 
-	err = c.DCDB.ExecSql("INSERT INTO config (first_load_blockchain, first_load_blockchain_url, auto_reload) VALUES (?, ?, ?)", firstLoad, url, 259200)
+	err = c.DCDB.ExecSql("INSERT INTO config (first_load_blockchain, first_load_blockchain_url, auto_reload) VALUES (?, ?, ?)", firstLoad, first_load_blockchain_url, 259200)
 	if err != nil {
 		log.Error("%v", utils.ErrInfo(err))
 		dropConfig()
@@ -166,22 +171,26 @@ func (c *Controller) Install() (string, error) {
 
 	log.Debug("GenerateFirstBlock", *utils.GenerateFirstBlock)
 
-	if _, err := os.Stat(*utils.FirstBlockDir + "/1block"); os.IsNotExist(err) {
+	if _, err := os.Stat(*utils.FirstBlockDir + "/1block"); len(*utils.FirstBlockDir) > 0 && os.IsNotExist(err) {
 
 		// If there is no key, this is the first run and the need to create them in the working directory.
 		if _, err := os.Stat(*utils.Dir + "/PrivateKey"); os.IsNotExist(err) {
 
 			if len(*utils.FirstBlockPublicKey) == 0 {
 				priv, pub := lib.GenKeys()
-				err := ioutil.WriteFile(*utils.Dir+"/PrivateKey", []byte(priv), 0644)
+				err := ioutil.WriteFile(*utils.Dir + "/PrivateKey", []byte(priv), 0644)
 				if err != nil {
 					log.Error("%v", utils.ErrInfo(err))
 				}
 				*utils.FirstBlockPublicKey = pub
 			}
+		}
+
+		if _, err := os.Stat(*utils.Dir + "/NodePrivateKey"); os.IsNotExist(err) {
 			if len(*utils.FirstBlockNodePublicKey) == 0 {
 				priv, pub := lib.GenKeys()
-				err := ioutil.WriteFile(*utils.Dir+"/NodePrivateKey", []byte(priv), 0644)
+				fmt.Println("WriteFile " + *utils.Dir + "/NodePrivateKey")
+				err := ioutil.WriteFile(*utils.Dir + "/NodePrivateKey", []byte(priv), 0644)
 				if err != nil {
 					log.Error("%v", utils.ErrInfo(err))
 				}
@@ -189,36 +198,36 @@ func (c *Controller) Install() (string, error) {
 			}
 		}
 
+		*utils.GenerateFirstBlock = 1
 		utils.FirstBlock(false)
-
+	}
 		log.Debug("1block")
 
-		NodePrivateKey, _ := ioutil.ReadFile(*utils.Dir + "/NodePrivateKey")
-		err = c.DCDB.ExecSql(`INSERT INTO my_node_keys (private_key, block_id) VALUES (?, ?)`, NodePrivateKey, 1)
-		if err != nil {
-			log.Error("%v", utils.ErrInfo(err))
-			dropConfig()
-			return "", utils.ErrInfo(err)
-		}
+	NodePrivateKey, _ := ioutil.ReadFile(*utils.Dir + "/NodePrivateKey")
+	NodePrivateKeyStr := strings.TrimSpace(string(NodePrivateKey))
+	npubkey := lib.PrivateToPublicHex(NodePrivateKeyStr)
+	err = c.DCDB.ExecSql(`INSERT INTO my_node_keys (private_key, public_key, block_id) VALUES (?, [hex], ?)`, NodePrivateKeyStr, npubkey, 1)
+	if err != nil {
+		log.Error("%v", utils.ErrInfo(err))
+		dropConfig()
+		return "", utils.ErrInfo(err)
+	}
+
+	if *utils.DltWalletId == 0 {
 		PrivateKey, _ := ioutil.ReadFile(*utils.Dir + "/PrivateKey")
 		PrivateHex, _ := hex.DecodeString(string(PrivateKey))
 		PublicKeyBytes2 := lib.PrivateToPublic(PrivateHex)
 		log.Debug("dlt_wallet_id %d", int64(lib.Address(PublicKeyBytes2)))
-
-		err = c.DCDB.ExecSql(`UPDATE config SET dlt_wallet_id = ?`, int64(lib.Address(PublicKeyBytes2)))
-		if err != nil {
-			log.Error("%v", utils.ErrInfo(err))
-			dropConfig()
-			return "", utils.ErrInfo(err)
-		}
-		/*		err = utils.DaylightRestart()
-				if err != nil {
-					log.Error("%v", utils.ErrInfo(err))
-					dropConfig()
-					return "", utils.ErrInfo(err)
-				}*/
-
+		*utils.DltWalletId = int64(lib.Address(PublicKeyBytes2))
 	}
+
+	err = c.DCDB.ExecSql(`UPDATE config SET dlt_wallet_id = ?`, *utils.DltWalletId)
+	if err != nil {
+		log.Error("%v", utils.ErrInfo(err))
+		dropConfig()
+		return "", utils.ErrInfo(err)
+	}
+
 	return `{"success":1}`, nil
 }
